requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-multipart
supabase
python-dotenv
google-genai
